Lock the sric tx buffer upon receiving a command from the host
I was previously worried that this could result in some kind of
deadlock, but after looking at the situation more closely I've
concluded that this isn't the case.  The host interface receives a
command frame and in the same interrupt context requests the lock from
the sric interface.  The lock is unavailable from the SRIC interface
in two situations:
 - When the interface is transmitting
 - When the interface has received a command frame

So the hostser and sric interfaces compete for locks on the sric
interface's tx buffer.  Whichever gets there first sets which one ends
up discarding its incoming frame (which is why, in the future, we will
need to do something about the gateway throwing away frames that the
host sends it to transmit).