Give up waiting for a response from a slave after 15 seconds
The normal retransmission stuff will happen using the token.  It's
really annoying if one has to reboot the BeagleBoard to restart
everything, so this patch adds an "error" callback to the SRIC
interface that is called when it gives up waiting for a response.  The
gateway takes this signal and resets itself when it receives this
message.