Add support for tokens to the sric state machine
This patch adds several new states to the SRIC state machine, which
permit waiting for the token.  A graphviz based state diagram of the
SRIC state machine will arrive soon.

This patch also adds the concept of a "token driver".  This set of
drivers will provide the interface defined in token-drv.h.  Included
is a dummy token driver that uses timeouts instead of the token ring
itself to emulate the token.