Add a token driver for the sric director to use
This generates the initial token, and passes it on when it gets it.
It currently injects a 2ms delay before retransmitting the token --
this is to prevent all of the chip's interrupt time being used for
token processing.

This 2ms latency will later be resolved by moving the pin interrupt
handling stuff to the mainloop.