Start reworking hostserial fsm:
Robs discovered that the current situation isn't going to work -
there are two independant resources, host transmit and host rx.
The occurence of one might invoke the other. Which leads to a call
to the FSM code occuring recursively, thus making it cry.

The fix is to seperate the two things - receiving resource and
transmitting resource. It seems two bools describing whether
we've received a frame or are transmitting a frame should be
sufficient.