Add some double buffering to sric:
This follows the same model as hostser - we have two buffers, an
index to indicate which one we're reading into, and whenever we get
a frame we update other pointers in the file to point at the new
full buffer.

A hack however - one of the pointers that has to be updated is in
the sric_if struct. Because of this, it can no longer be const. Eww.