Use pointer to current incoming sric data buffer:
I believe using a pointer rather than referencing the two dimensional
array is going to be more efficient in terms of cycles used. Given
that the array/pointer is referenced three times, and this code adds
an extra calculation to the longest rx intr path, I think it's
probably neutral inre the number of cycles saved.