Further shorten rx_fsm's interrupt operation:
We now take into account the fact that there are only two buffers
(and eliminate rxbuf_idx), perform a comparison to see which
buffer we were working on, and select the next one appropriately.

This eliminates a costly address-of operation, which takes something
like 12 cycles. In comparison, this (seems to) takes four.