Fix stop-bit clobbering fault by sending more padding bytes:
See the srobo-devel mailing list on 09/12/2010 for more details
on precisely what the problem is; in short, the token was being
advanced around the network before a device had sent it's stop
bit onto the bus, resulting in a data collision with the next
device (if it transmitted).

This sends two padding bytes after each frame, and disables
transmission after the first (see list for why this works).