Enable rxing from sric bus immediately after disabling txing:
Previously in tokenless mode, we had the following events:
1) Gateway sends command (say, addr-assign)
2) Gateway sends padding byte
3) Device sends ack/response for addr-addign

However, with the recent fix in 865660b06a0cc201, we instead have
the following events:
1) Gateway sends command (say, addr-assign)
2) Gateway sends 1st padding byte
3) Device sends ack/response for addr-addign
4) Gateway sends 2nd padding byte

This means that, due to the extra padding byte, the gateway misses
the start-of-frame from the device. This patch enables receiving
immediately after we send the first padding byte (and disable
txing), so there's extra time for the devices SOF to turn up in.

Fundementally tokenless mode is unsynchronised, so there's no
pretty way to fix this.