Correctly deal with RESPOND_NOW flag in sric-client return codes
Specifically, if the command uses the RESPOND_NOW flag, it gets
interpreted as an error code, and the invoke handler exits ahead
of time without filling out the src/dst/sz fields of the frame.

Also, the RESPOND_NOW flag should be masked out of the frame size
that gets sent.