Insert a 1ms delay when responding to system commands:
This is to ensure that there's sufficient delay between the receipt
of a system message that is going to be replied to, and the device
generating it finishing its transmission.

During enumeration, there's no guarentee that the bus director and
client with a token are not going to transmit at the same time. As
it happens, there's a very short period where a device receives a
message and sends a response, which races with the padding byte the
bus director sends. If the start bit and stop bit of each
respectively collide, we get in trouble.

SRIC retransmission is the more appropriate fix, however it's possible
for sric enumeration to never make progress due to this race.