Adjust sric-gws handling of incoming messages:
If it's to be forwarded to sric, then consider whether to discard it because
sric is already in use. However if it's destined for the gw interface, then
never discard, we want to be able to communicate with the gw even if sric
has hung up.

Also, remove a bunch of hostser_rx_done calls to the end of the inhost_fsm,
which achieves exactly the same things, just in one place.