If multiple tokens arrive, don't schedule one million tasks
Detected by Rich - if you plug/deplug the in port to the power board,
then it generates a ton of token (expected) and hangs up (not expected).
This is because it receives a bunch of token intrs, and schedules one
task for each. Fix this by making the token emit callback only schedule
itself once.