Add a system command for the discovery of firmware git versions
I think it will be mildly useful to be able to "ask" a board more
about what firmware version is has within it.  Yes, we have the
flash430 library, but this works with version numbers that we generate
ourselves, and so can result in erroneous identification of firmware.

This patch adds a SRIC system command (i.e. it will be common to all
boards that use libsric) that allows an autogenerated "versionbuf" to
be read.  The versionbuf contains the git commit hashes of the main
firmware repository, as well as the submodules that it contains.

The cost for storing this data is quite low, at only 51 bytes of flash
for the motor controller.  I've truncated the git hashes that are
stored within the versionbuf to the first 5 bytes, which should be
enough data to uniquely identify commits within repositories.

I will modify pyenv to read the versionbufs of all the boards upon
boot and write the results to a file on the USB key.  Hopefully this
will provide some useful information when debugging things.