Fix maximum length check of response frames, allowing >57 byte payloads
The sric-if calls a function to get the response to a received frame.
This function returns the number of bytes that it put in the transmit
buffer to send as a response.  sric-if compared this value against
MAX_PAYLOAD.  This is wrong, as the response includes the headers
too.  This meant that responses that had a payload of more than 57
bytes would be thrown away immediately after the transmit buffer had
been filled with them.

This patch fixes the value that the response length is compared
against.  It adds a new #define called "MAX_FRAME_LEN", which includes
the checksum on the end of the frame.  Thus, the comparison must be
against MAX_FRAME_LEN - 2.

The same applied to the gateway's length checking.  This patch fixes
that too.

With this patch, I can successfully transmit 64-byte payloads through
the bus.