Hash out the shape of the gateway sric interface
After a lot of beating, it's now apparent that the gateway dev will
need it's own retransmission logic. Uuugh, code replication.

Anyway - the interface is the same as sric (sric_if_t), and just sends
a frame to hostser if retransmission is not desired. If it /is/
desired, we stash a copy of the frame, then after 10ms or so time out
and send it again. Handling the ack is yet to be done.