Fix power board frame retransmission situation
We were triggering some races (based on the fact that sricd won't
respond to frames from the power board if it's waiting for an ack
from something on the bus). So, it's not good for the gateway to
block all incoming traffic while it's in retransmit mode.

We don't have enough space for an extra 70 byte buffer, so instead
use a 12 byte one and rely on the fact that the power board won't
send anything bigger in a note.

Yes, hacks.