Add token-loss monitor.
After the first token release, this checks 4 times a second to see if
we've seen the token recently. If the token hasn't been seen, it is
regenerated.

This is enough time for 10 boards to make a 70 byte transmission each,
several times. I've tested and observed this working (loss then
regeneration)

The token really shouldn't be lost, ever, anyway. However it's
happening a little too often while we're testing boards, and this
isn't going to hurt anything.